Atualização do algoritmo de geração de veículos.

+++ core/src/br/com/chickenroad/screens/util/Util.java
@@ -6,8 +6,8 @@ public class Util {
 	
 	private static Random random = new Random();
 	
-	public static int getRandomPosition(float min, float max){
-		return random.nextInt((int)max - (int)min) + (int)min;
+	public static float getRandomPosition(float min, float max){
+		return random.nextInt((int)max - (int)min) + min;
 	}
 	public static float getRandomPosition(float min){
 		return random.nextFloat() + min;
